refactor: S3 이미지 업로드 클래스 이름과 패키지 변경
- 클래스 이름 변경 : AwsS3Service -> AwsS3ImageUploader
- 패키지 변경 : common.util -> common.aws
