+++ src/main/java/com/tasty/masiottae/common/aws/AwsS3ImageUploader.java
@@ -1,7 +1,6 @@
-package com.tasty.masiottae.common.util;
+package com.tasty.masiottae.common.aws;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
@@ -16,7 +15,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
-public class AwsS3Service {
+public class AwsS3ImageUploader {
 
     private final AmazonS3 s3Client;
     @Value("${cloud.aws.s3.bucket}")
